requirements.txt
faker
pyttsx3
requests
python-dotenv
ollama
tqdm